+++ fs/src/fs/org/jnode/fs/hfsplus/SuperBlock.java
@@ -17,14 +17,13 @@
  * along with this library; If not, write to the Free Software Foundation, Inc., 
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
- 
+
 package org.jnode.fs.hfsplus;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.Calendar;
 import java.util.Date;
-
 import org.apache.log4j.Level;
 import org.apache.log4j.Logger;
 import org.jnode.fs.FileSystemException;
@@ -35,9 +34,8 @@ import org.jnode.util.NumberUtils;
 
 /**
  * HFS+ volume header definition.
- * 
+ *
  * @author Fabien Lesire
- * 
  */
 public class SuperBlock extends HfsPlusObject {
 
@@ -58,20 +56,23 @@ public class SuperBlock extends HfsPlusObject {
 
     private final Logger log = Logger.getLogger(getClass());
 
-    /** Volume header data length */
+    /**
+     * Volume header data length
+     */
     public static final int SUPERBLOCK_LENGTH = 1024;
 
-    /** Data bytes array that contains volume header information */
+    /**
+     * Data bytes array that contains volume header information
+     */
     private byte[] data;
 
     /**
      * Create the volume header and load information for the file system passed
      * as parameter.
-     * 
+     *
      * @param fs The file system contains HFS+ partition.
-     * 
      * @throws FileSystemException If magic number (0X482B) is incorrect or not
-     *             available.
+     *                             available.
      */
     public SuperBlock(final HfsPlusFileSystem fs, boolean create) throws FileSystemException {
         super(fs);
@@ -88,7 +89,7 @@ public class SuperBlock extends HfsPlusObject {
                 System.arraycopy(b.array(), 0, data, 0, SUPERBLOCK_LENGTH);
                 if (getMagic() != HFSPLUS_SUPER_MAGIC && getMagic() != HFSX_SUPER_MAGIC) {
                     throw new FileSystemException("Not hfs+ volume header (" + getMagic() +
-                            ": bad magic)");
+                        ": bad magic)");
                 }
 
             }
@@ -99,13 +100,13 @@ public class SuperBlock extends HfsPlusObject {
 
     /**
      * Create a new volume header.
-     * 
+     *
      * @param params File system format parameters.
      * @throws IOException
      */
     public void create(HFSPlusParams params) throws IOException {
         log.info("Create new HFS+ volume header (" + params.getVolumeName() +
-                ") with block size of " + params.getBlockSize() + " bytes.");
+            ") with block size of " + params.getBlockSize() + " bytes.");
         int burnedBlocksBeforeVH = 0;
         int burnedBlocksAfterAltVH = 0;
         /*
@@ -146,8 +147,8 @@ public class SuperBlock extends HfsPlusObject {
         int startBlock = 1 + burnedBlocksBeforeVH;
         int blockCount = (int) bitmapBlocks;
         HfsPlusForkData forkdata =
-                new HfsPlusForkData(allocationClumpSize, (int) allocationClumpSize,
-                        (int) bitmapBlocks);
+            new HfsPlusForkData(allocationClumpSize, (int) allocationClumpSize,
+                (int) bitmapBlocks);
         ExtentDescriptor desc = new ExtentDescriptor(startBlock, blockCount);
         forkdata.addDescriptor(0, desc);
         forkdata.write(data, 112);
@@ -166,8 +167,8 @@ public class SuperBlock extends HfsPlusObject {
         // Extent B-Tree initialization
         log.info("Init extent file.");
         forkdata =
-                new HfsPlusForkData(params.getExtentClumpSize(), params.getExtentClumpSize(),
-                        (params.getExtentClumpSize() / blockSize));
+            new HfsPlusForkData(params.getExtentClumpSize(), params.getExtentClumpSize(),
+                (params.getExtentClumpSize() / blockSize));
         desc = new ExtentDescriptor(nextBlock, forkdata.getTotalBlocks());
         forkdata.addDescriptor(0, desc);
         forkdata.write(data, 192);
@@ -177,8 +178,8 @@ public class SuperBlock extends HfsPlusObject {
         log.info("Init catalog file.");
         int totalBlocks = params.getCatalogClumpSize() / blockSize;
         forkdata =
-                new HfsPlusForkData(params.getCatalogClumpSize(), params.getCatalogClumpSize(),
-                        totalBlocks);
+            new HfsPlusForkData(params.getCatalogClumpSize(), params.getCatalogClumpSize(),
+                totalBlocks);
         desc = new ExtentDescriptor(nextBlock, totalBlocks);
         forkdata.addDescriptor(0, desc);
         forkdata.write(data, 272);
@@ -186,12 +187,12 @@ public class SuperBlock extends HfsPlusObject {
 
         this.setFreeBlocks(this.getFreeBlocks() - (int) blockUsed);
         this.setNextAllocation((int) blockUsed - 1 - burnedBlocksAfterAltVH + 10 *
-                (this.getCatalogFile().getClumpSize() / this.getBlockSize()));
+            (this.getCatalogFile().getClumpSize() / this.getBlockSize()));
     }
 
     /**
      * Calculate the number of blocks needed for bitmap.
-     * 
+     *
      * @param totalBlocks Total of blocks found in the device.
      * @return the number of blocks.
      * @throws IOException
@@ -209,7 +210,7 @@ public class SuperBlock extends HfsPlusObject {
     // Getters/setters
 
     public final int getMagic() {
-        return BigEndian.getInt16(data, 0);
+        return BigEndian.getUInt16(data, 0);
     }
 
     public final void setMagic(final int value) {
@@ -218,7 +219,7 @@ public class SuperBlock extends HfsPlusObject {
 
     //
     public final int getVersion() {
-        return BigEndian.getInt16(data, 2);
+        return BigEndian.getUInt16(data, 2);
     }
 
     public final void setVersion(final int value) {
@@ -226,18 +227,18 @@ public class SuperBlock extends HfsPlusObject {
     }
 
     //
-    public final int getAttributes() {
+    public final long getAttributes() {
 
-        return BigEndian.getInt32(data, 4);
+        return BigEndian.getUInt32(data, 4);
     }
 
     public final void setAttribute(final int attributeMaskBit) {
-        BigEndian.setInt32(data, 4, getAttributes() | (1 << attributeMaskBit));
+        BigEndian.setInt32(data, 4, (int) (getAttributes() | (1 << attributeMaskBit)));
     }
 
     //
-    public final int getLastMountedVersion() {
-        return BigEndian.getInt32(data, 8);
+    public final long getLastMountedVersion() {
+        return BigEndian.getUInt32(data, 8);
     }
 
     public final void setLastMountedVersion(final int value) {
@@ -245,8 +246,8 @@ public class SuperBlock extends HfsPlusObject {
     }
 
     //
-    public final int getJournalInfoBlock() {
-        return BigEndian.getInt32(data, 12);
+    public final long getJournalInfoBlock() {
+        return BigEndian.getUInt32(data, 12);
     }
 
     public final void setJournalInfoBlock(final long value) {
@@ -287,8 +288,8 @@ public class SuperBlock extends HfsPlusObject {
     }
 
     //
-    public final int getFileCount() {
-        return BigEndian.getInt32(data, 32);
+    public final long getFileCount() {
+        return BigEndian.getUInt32(data, 32);
     }
 
     public final void setFileCount(final int value) {
@@ -296,12 +297,12 @@ public class SuperBlock extends HfsPlusObject {
     }
 
     //
-    public final int getFolderCount() {
-        return BigEndian.getInt32(data, 36);
+    public final long getFolderCount() {
+        return BigEndian.getUInt32(data, 36);
     }
 
-    public final void setFolderCount(final int value) {
-        BigEndian.setInt32(data, 36, value);
+    public final void setFolderCount(final long value) {
+        BigEndian.setInt32(data, 36, (int) value);
     }
 
     //
@@ -314,8 +315,8 @@ public class SuperBlock extends HfsPlusObject {
     }
 
     //
-    public final int getTotalBlocks() {
-        return BigEndian.getInt32(data, 44);
+    public final long getTotalBlocks() {
+        return BigEndian.getUInt32(data, 44);
     }
 
     public final void setTotalBlocks(final int value) {
@@ -323,17 +324,17 @@ public class SuperBlock extends HfsPlusObject {
     }
 
     //
-    public final int getFreeBlocks() {
-        return BigEndian.getInt32(data, 48);
+    public final long getFreeBlocks() {
+        return BigEndian.getUInt32(data, 48);
     }
 
-    public final void setFreeBlocks(final int value) {
-        BigEndian.setInt32(data, 48, value);
+    public final void setFreeBlocks(final long value) {
+        BigEndian.setInt32(data, 48, (int) value);
     }
 
     //
-    public final int getNextAllocation() {
-        return BigEndian.getInt32(data, 52);
+    public final long getNextAllocation() {
+        return BigEndian.getUInt32(data, 52);
     }
 
     public final void setNextAllocation(final int value) {
@@ -341,31 +342,31 @@ public class SuperBlock extends HfsPlusObject {
     }
 
     public final long getRsrcClumpSize() {
-        return BigEndian.getInt32(data, 56);
+        return BigEndian.getUInt32(data, 56);
     }
 
     public final void setRsrcClumpSize(final int value) {
         BigEndian.setInt32(data, 56, value);
     }
 
-    public final int getDataClumpSize() {
-        return BigEndian.getInt32(data, 60);
+    public final long getDataClumpSize() {
+        return BigEndian.getUInt32(data, 60);
     }
 
     public final void setDataClumpSize(final int value) {
         BigEndian.setInt32(data, 60, value);
     }
 
-    public final int getNextCatalogId() {
-        return BigEndian.getInt32(data, 64);
+    public final long getNextCatalogId() {
+        return BigEndian.getUInt32(data, 64);
     }
 
-    public final void setNextCatalogId(final int value) {
-        BigEndian.setInt32(data, 64, value);
+    public final void setNextCatalogId(final long value) {
+        BigEndian.setInt32(data, 64, (int) value);
     }
 
-    public final int getWriteCount() {
-        return BigEndian.getInt32(data, 68);
+    public final long getWriteCount() {
+        return BigEndian.getUInt32(data, 68);
     }
 
     public final void setWriteCount(final int value) {
@@ -408,30 +409,29 @@ public class SuperBlock extends HfsPlusObject {
 
     /**
      * Get string representation of attribute.
-     * 
+     *
      * @return the string representation
      */
     public final String getAttributesAsString() {
         return ((isAttribute(HFSPLUS_VOL_UNMNT_BIT)) ? " kHFSVolumeUnmountedBit" : "") +
-                ((isAttribute(HFSPLUS_VOL_INCNSTNT_BIT)) ? " kHFSBootVolumeInconsistentBit" : "") +
-                ((isAttribute(HFSPLUS_VOL_JOURNALED_BIT)) ? " kHFSVolumeJournaledBit" : "");
+            ((isAttribute(HFSPLUS_VOL_INCNSTNT_BIT)) ? " kHFSBootVolumeInconsistentBit" : "") +
+            ((isAttribute(HFSPLUS_VOL_JOURNALED_BIT)) ? " kHFSVolumeJournaledBit" : "");
     }
 
     /**
      * Check if the corresponding attribute corresponding is set.
-     * 
+     *
      * @param maskBit Bit position of the attribute. See constants.
-     * 
      * @return {@code true} if attribute is set.
      */
     public final boolean isAttribute(final int maskBit) {
         return (((getAttributes() >> maskBit) & 0x1) != 0);
     }
-    
-    public void incrementFolderCount(){
-    	this.setFolderCount(this.getFolderCount() + 1);
+
+    public void incrementFolderCount() {
+        this.setFolderCount(this.getFolderCount() + 1);
     }
-    		
+
 
     public byte[] getBytes() {
         return data;
@@ -446,16 +446,16 @@ public class SuperBlock extends HfsPlusObject {
         buffer.append("Magic: 0x").append(NumberUtils.hex(getMagic(), 4)).append("\n");
         buffer.append("Version: ").append(getVersion()).append("\n").append("\n");
         buffer.append("Attributes: ").append(getAttributesAsString()).append(" (").append(
-                getAttributes()).append(")").append("\n").append("\n");
+            getAttributes()).append(")").append("\n").append("\n");
         buffer.append("Create date: ").append(
-                HfsUtils.printDate(getCreateDate(), "EEE MMM d HH:mm:ss yyyy")).append("\n");
+            HfsUtils.printDate(getCreateDate(), "EEE MMM d HH:mm:ss yyyy")).append("\n");
         buffer.append("Modify date: ").append(
-                HfsUtils.printDate(getModifyDate(), "EEE MMM d HH:mm:ss yyyy")).append("\n");
+            HfsUtils.printDate(getModifyDate(), "EEE MMM d HH:mm:ss yyyy")).append("\n");
         buffer.append("Backup date: ").append(
-                HfsUtils.printDate(getBackupDate(), "EEE MMM d HH:mm:ss yyyy")).append("\n");
+            HfsUtils.printDate(getBackupDate(), "EEE MMM d HH:mm:ss yyyy")).append("\n");
         buffer.append("Checked date: ").append(
-                HfsUtils.printDate(getCheckedDate(), "EEE MMM d HH:mm:ss yyyy")).append("\n")
-                .append("\n");
+            HfsUtils.printDate(getCheckedDate(), "EEE MMM d HH:mm:ss yyyy")).append("\n")
+            .append("\n");
         buffer.append("File count: ").append(getFileCount()).append("\n");
         buffer.append("Folder count: ").append(getFolderCount()).append("\n").append("\n");
         buffer.append("Block size: ").append(getBlockSize()).append("\n");
