+++ fs/src/fs/org/jnode/fs/hfsplus/tree/NodeDescriptor.java
@@ -17,7 +17,7 @@
  * along with this library; If not, write to the Free Software Foundation, Inc., 
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
- 
+
 package org.jnode.fs.hfsplus.tree;
 
 import org.jnode.util.BigEndian;
@@ -28,27 +28,39 @@ public class NodeDescriptor {
     public static final int BT_HEADER_NODE = 1;
     public static final int BT_MAP_NODE = 2;
 
-    /** The size of the node descriptor. */
+    /**
+     * The size of the node descriptor.
+     */
     public static final int BT_NODE_DESCRIPTOR_LENGTH = 14;
 
-    /** The number of the next node. */
-    private int fLink;
+    /**
+     * The number of the next node.
+     */
+    private long fLink;
 
-    /** The number of the previous node. */
-    private int bLink;
+    /**
+     * The number of the previous node.
+     */
+    private long bLink;
 
-    /** The type of the node. */
+    /**
+     * The type of the node.
+     */
     private int kind;
 
-    /** The depth of this node in the B-Tree. */
+    /**
+     * The depth of this node in the B-Tree.
+     */
     private int height;
 
-    /** The number of records in this node. */
+    /**
+     * The number of records in this node.
+     */
     private int numRecords;
 
     /**
      * Creates a new node descriptor.
-     * 
+     *
      * @param fLink
      * @param bLink
      * @param kind
@@ -65,28 +77,27 @@ public class NodeDescriptor {
 
     /**
      * Creates node descriptor from existing data.
-     * 
-     * @param src byte array contains node descriptor data.
+     *
+     * @param src    byte array contains node descriptor data.
      * @param offset start of node descriptor data.
      */
     public NodeDescriptor(final byte[] src, int offset) {
         byte[] data = new byte[BT_NODE_DESCRIPTOR_LENGTH];
         System.arraycopy(src, offset, data, 0, BT_NODE_DESCRIPTOR_LENGTH);
-        fLink = BigEndian.getInt32(data, 0);
-        bLink = BigEndian.getInt32(data, 4);
+        fLink = BigEndian.getUInt32(data, 0);
+        bLink = BigEndian.getUInt32(data, 4);
         kind = BigEndian.getInt8(data, 8);
-        height = BigEndian.getInt8(data, 9);
-        numRecords = BigEndian.getInt16(data, 10);
+        height = BigEndian.getUInt8(data, 9);
+        numRecords = BigEndian.getUInt16(data, 10);
     }
 
     /**
-     * 
      * @return the descriptor rendered as bytes
      */
     public byte[] getBytes() {
         byte[] data = new byte[BT_NODE_DESCRIPTOR_LENGTH];
-        BigEndian.setInt32(data, 0, fLink);
-        BigEndian.setInt32(data, 4, bLink);
+        BigEndian.setInt32(data, 0, (int) fLink);
+        BigEndian.setInt32(data, 4, (int) bLink);
         BigEndian.setInt8(data, 8, kind);
         BigEndian.setInt8(data, 9, height);
         BigEndian.setInt16(data, 10, numRecords);
@@ -95,14 +106,14 @@ public class NodeDescriptor {
 
     public final String toString() {
         return ("FLink:  " + getFLink() + "\n" + "BLink:  " + getBLink() + "\n" + "Kind:   " +
-                getKind() + "\n" + "height: " + getHeight() + "\n" + "#rec:   " + getNumRecords() + "\n");
+            getKind() + "\n" + "height: " + getHeight() + "\n" + "#rec:   " + getNumRecords() + "\n");
     }
 
-    public int getFLink() {
+    public long getFLink() {
         return fLink;
     }
 
-    public int getBLink() {
+    public long getBLink() {
         return bLink;
     }
